pyproject.toml
[project]
name = "amadeus-fastapi"
version = "0.1.0"
description = "FastAPI application for Amadeus APIs"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi",
  "uvicorn",
  "python-dotenv",
  "httpx",
  "requests",
  "pydantic",
  "amadeus",
  "pydotenv",
]
